google/falco/Kconfig: Move select MAINBOARD_DO_NATIVE_VGA_INIT
Move the default select of "Use native graphics initialization" for
Falco to the ChromeOS section as SeaBIOS (default payload) requires a
vBIOS and takes twice as long to load with this option enabled. For the
same reasons, this option shouldn't be enabled by default (def_bool y).

adapted from identical upstream commit for google/peppy,
6d2ca021014aa967fc0be83f2f0d76e423db2037

Change-Id: I033d0a04fe22ff70e2162a61a673caf813fab27a
Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>