google/cyan: set touchscreen GPIO to non_maskable
a recent commit which changed the touchpad/touchscreen interrupts
from edge to level triggered also marked them as maskable.  This
not only broke the touchpad functionality, but caused issues with
the touchpad as well.  Revert the touchpad to being non_maskable
for all cyan variants with a touchscreen.

Test: boot GalliumOS on google/cyan with a range of kernel versions
(4.15.18, 4.16.13, 4.17.x, 4.18.x) and verify touchscreen functional,
touchpad working properly (not jittery)

Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>