x86/smbios_defaults: Set default system manufacturer for ChromeOS devices
Currently, many Linux drivers use DMI quirks to identify ChromeOS devices and
handle them accordingly: namely they look for the SMBIOS system manufactuer
to be "GOOGLE" or "Google", and the bios-vendor to be coreboot. Historically
this was consistently the case, but recent model ChromeOS devices allow the
OEM to set the mainboard manufacturer, which is also the default system
manufacturer. This breaks many DMI quirks, notably ones used by SOF (sound
open firmware) for audio.

To fix this, set the default system manufactuer for ChromeOS devices to
"Google" conditional on CONFIG_MAINBOARD_HAS_CHROMEOS, leaving the OEM
customization in place for the mainboard manufacturer.

Test: build/boot google/bloog with Linux 5.16, verify SOF drivers
correctly detect device as a Chromebook and load the appropriate
audio firmware.

Change-Id: I9de17fa12689ab4e627b995818aa3d2653102b04
Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>