payloads/tianocore: Fix up Kconfig repo, bootsplash
Commit 0884f210 [payloads/tianocore: Rework Makefile] added a "custom"
option for the edk2 payload, but neglected to clear the default edk2
repo and tag/branch fields, leaving the defaults for the UEFIPAYLOAD
option. Clear these to make it ovbious they should be set by the user.

It also removed the Kconfig guarding the use of a custom bootsplash
image, leaving the default "bootsplash.bmp" which does not exist,
causing a silent failure when building. Add a BMP of the coreboot
logo as the default bootsplash and adjust the path/filename.
Correct the help text since it was not adjusted when the "use"
Kconfig was removed in 0084f210.

Change-Id: I1b6af815564fc02eb93c4727d15963361095cc75
Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>