Fix Byt CB Bootup on SeaBIOS when using coreboot
This patch helps fix the internal eMMC card when booting from it with
the MrChromebox custom coreboot distro. If this patch isn't applied,
the OS cannot boot off of it, and if you boot from USB, then you will
get spamming errors in dmesg. This also doesn't show on Windows PE if
you do not apply this, and it will be very slow to boot, or to even
use. Note that this SMM variable is currently only in the MrChromebox
coreboot distro at the moment, and is used by his EDK2 distro as well.
Also the line added in this patch should not be added to CBFS
booting, so we can assume that the next payload woll do the same thing,
like if we chainload SeaBIOS to EDK2. Also this is a dirty patch and
should not be sent to mainline SeaBIOS and people hope for it to be
removed with a coreboot fix anyways, when it isn't needed in EDK2
as well to boot or use. Tested only on GOOGLE Enguarde at the moment.
MrChromebox = Matt DeVillier
Byt = Intel Bay Trail
CB = ChromeBook (and other Chrome OS Devices like Chromebox)

Signed-off-by: Christopher Lentocha <christopherericlentocha@gmail.com>