soc/amd/stoneyridge: Update SMU fw2 name in fw.cfg
Update the filename for the PSP_SMUFW2_SUB1_FILE to use the compressed
and signed version (.csbin) rather than the uncompression + signed
version (.sbin) as they are the same file, in order to be consistent
with the other SMU firmware files. This will also facilitate dropping
the duplicate files in an upcoming update to the amd_blobs repo and
updating the SMU files (all .csbin).

md5sum comparison:
7ea83292c452b637b56d3991310c4c1d  SmuFirmware2_prod_ST_FN.csbin
7ea83292c452b637b56d3991310c4c1d  SmuFirmware2_prod_ST_FN.sbin

TEST=build/boot google/kahlee (liara)

Change-Id: I10fa8e949ab589d315862c06b4125c902520cbbc
Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>