acpi/acpigen: introduce acpigen_pop_len_compress
acpigen_pop_len will always write a 3 byte PkgLength to the 3 bytes
reserved by acpigen_write_len_f. In contrast to that, the newly
introduced acpigen_pop_len_compress writes a 1-3 byte PkgLength
depending on the number of bytes in between. When less than 3 bytes are
written for PkgLength, the data afterwards will be moved forward by
number of reserved bytes that aren't needed for the PkgLength.

Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Change-Id: Ib897b08a05a7cdc52902d51364246c260ea1f206